requirements.txt
# Auto Channel Bot — Python Dependencies
# Render installs these automatically on every deploy

python-telegram-bot==21.3
feedparser==6.0.11
httpx==0.27.0
beautifulsoup4==4.12.3
lxml==5.2.2
APScheduler==3.10.4
openai==1.30.0
praw==7.7.1
tweepy==4.14.0
Pillow==10.3.0
